make_html: Copy template instead of setting intermediates_dir
If the intermediates_dir option is used, making reports within
vignettes fails with spurious duplicated sections.

There are various reports of it not working very well, so don't use it.
Instead, copy the template to the final directory, and now the defaults
for knit_root_dir / intermediates_dir will do the right thing.